// packages/core/src/squads/generated/types/configAction.ts
export type ConfigAction =
  | { __kind: "AddMember"; newMember: Member }
  | { __kind: "RemoveMember"; oldMember: Address }
  | { __kind: "ChangeThreshold"; newThreshold: number }
  | { __kind: "SetTimeLock"; newTimeLock: number }
  | {
      __kind: "AddSpendingLimit";
      /** Key that is used to seed the SpendingLimit PDA. */
      createKey: Address;
      /** The index of the vault that the spending limit is for. */
      vaultIndex: number;
      /** The token mint the spending limit is for. */
      mint: Address;
      /**
       * The amount of tokens that can be spent in a period.
       * This amount is in decimals of the mint,
       * so 1 SOL would be `1_000_000_000` and 1 USDC would be `1_000_000`.
       */
      amount: bigint;
      /**
       * The reset period of the spending limit.
       * When it passes, the remaining amount is reset, unless it's `Period::OneTime`.
       */
      period: Period;
      /**
       * Members of the multisig that can use the spending limit.
       * In case a member is removed from the multisig, the spending limit will remain existent
       * (until explicitly deleted), but the removed member will not be able to use it anymore.
       */
      members: Array<Address>;
      /**
       * The destination addresses the spending limit is allowed to sent funds to.
       * If empty, funds can be sent to any address.
       */
      destinations: Array<Address>;
    }
  | { __kind: "RemoveSpendingLimit"; spendingLimit: Address }
  | { __kind: "SetRentCollector"; newRentCollector: Option<Address> };